# web/storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'Payment Gateways API Microservice'
  description: 'Payment Gateways API Microservice, Version v1'
  contact:
    name: 'Support Team'
    email: support@ultainfinity.com
  version: v1
servers:
  -
    url: 'http://localhost:8100/v1'
    description: 'Payment Gateways API Microservice, Version v1'
paths:
...
  /admin/payment-system:
    get:
      tags:
        - 'Admin / Payment-System'
      summary: 'Display list of all payment system'
      description: 'Display list of all payment gateway settings'
      operationId: 10e1b21d7640a333265c70d27fc6e9d4
      parameters:
        -
          name: limit
          in: query
          description: 'Count of orders in response'
          required: false
          schema:
            type: integer
            default: 20
        -
          name: page
          in: query
          description: 'Page of list'
          required: false
          schema:
            type: integer
            default: 1
      responses:
        '200':
          description: Success
      security:
        -
          default:
            - ManagerRead
            - User
            - ManagerWrite
      x-auth-type: 'Application & Application User'
      x-throttling-tier: Unlimited
      x-wso2-application-security:
        security-types:
          - oauth2
        optional: 'false'
    post:
      tags:
        - 'Admin / Payment-System'
      summary: 'Method to add new payment system'
      description: 'method to add new payment system'
      operationId: f4cd0f0c1be80beedbc21c8ce1cd18da
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                name:
                  description: 'Name of the payment system'
                  type: string
                gateway:
                  description: 'The payment system gateway'
                  type: string
                description:
                  description: 'details of the payment system'
                  type: string
                new_status:
                  description: 'Currently in use settings'
                  type: integer
              type: object
      responses:
        '200':
          description: Success
      security:
        -
          default:
            - ManagerRead
            - User
            - ManagerWrite
      x-auth-type: 'Application & Application User'
      x-throttling-tier: Unlimited
      x-wso2-application-security:
        security-types:
          - oauth2
        optional: 'false'
  '/admin/{id}/payment-system':
    get:
      tags:
        - 'Admin / Payment-System'
      summary: 'Display payment system details'
      description: 'show payment system details'
      operationId: 6ce8fe6ba19a457507a1ab492bf18623
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                id:
                  description: 'primary key to the record'
                  type: string
              type: object
      responses:
        '200':
          description: Success
      security:
        -
          default:
            - ManagerRead
            - User
            - ManagerWrite
      x-auth-type: 'Application & Application User'
      x-throttling-tier: Unlimited
      x-wso2-application-security:
        security-types:
          - oauth2
        optional: 'false'
    put:
      tags:
        - 'Admin / Payment-System'
      summary: 'Method to update payment System'
      description: 'method to update payment system'
      operationId: 5bc074f41873b1f465db7453a6bfb40d
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                id:
                  description: 'primary key to the record'
                  type: string
                name:
                  description: 'Name of the payment system'
                  type: string
                gateway:
                  description: 'The payment system gateway'
                  type: string
                description:
                  description: 'details of the payment system'
                  type: string
                new_status:
                  description: 'Currently in use settings'
                  type: integer
              type: object
      responses:
        '200':
          description: Success
      security:
        -
          default:
            - ManagerRead
            - User
            - ManagerWrite
      x-auth-type: 'Application & Application User'
      x-throttling-tier: Unlimited
      x-wso2-application-security:
        security-types:
          - oauth2
        optional: 'false'
    delete:
      tags:
        - 'Admin / Payment-System'
      summary: 'Method to delete payment system'
      description: 'method to delete payment gateway settings'
      operationId: fe110cd177a7906b682da2354192f88a
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                id:
                  description: 'payment system id'
                  type: string
              type: object
      responses:
        '200':
          description: Success
      security:
        -
          default:
            - ManagerRead
            - User
            - ManagerWrite
      x-auth-type: 'Application & Application User'
      x-throttling-tier: Unlimited
      x-wso2-application-security:
        security-types:
          - oauth2
        optional: 'false'
...
components:
  securitySchemes:
    default:
      type: oauth2
      description: 'Auth Scheme'
      flows:
        implicit:
          authorizationUrl: 'https://sumraid.com/oauth2'
          scopes:
            ManagerRead: 'Manager can read'
            User: 'User access'
            ManagerWrite: 'Manager can write'